cdk/requirements.txt
Flask==2.0.1
boto3==1.18.0
aws-cdk.core==1.130.0
aws-cdk.aws_lambda==1.130.0
aws-cdk.aws_apigateway==1.130.0
aws-cdk.aws_dynamodb==1.130.0